refactor(administration): split extension tooling scripts into focused modules
The opt-in extension tooling CLI had grown a few very large files: check.ts
(1505 lines, including a ~530-line checkExtensions god function), setup.ts
(1280) and report.ts (1079). Decompose them into cohesive, single-
responsibility modules, keeping each original file as a re-export barrel so
the public import surface and the Jest specs stay unchanged.

- check.ts (1505 -> 417): execution pipeline -> check-run, source discovery
  and output parsing -> check-parsing, vue-tsc program -> check-typescript-
  program, shared primitives -> check-pipeline, data model -> check-types
- setup.ts (1280 -> 358): -> setup-configs, setup-bridge, setup-discovery,
  setup-gitignore, setup-context
- report.ts (1079 -> 21 barrel): -> report-check, report-setup,
  report-guidance, report-summary
- probe.ts: extract the child-process runner -> probe-command
- consolidate the duplicated readEslintMajorVersion into shared.ts

No behavioural change: all user-facing output and the NUL-byte key separators
are preserved. The extension tooling Jest suite (23 unit + 9 e2e) and ESLint
pass unchanged.
