+++ src/Administration/Resources/app/administration/scripts/extensionTooling/setup-gitignore.ts
@@ -1,0 +1,81 @@
+/**
+ * @sw-package framework
+ *
+ * Ownership-consistent management of a fenced block in the project .gitignore:
+ * own the block, never rewrite user lines, respect a user's deletion of it, and
+ * stand down when the entries are already covered or after --no-gitignore.
+ */
+
+import fs from 'fs';
+import path from 'path';
+import { record, toManifestState } from './setup-context';
+import type { GeneratorContext } from './setup-context';
+import { MANAGED_BLOCK_BEGIN, writeManagedBlock } from './shared';
+import type { ManifestFileState } from './shared';
+
+const GITIGNORE_ENTRIES = [
+    '/tsconfig.json',
+    '/eslint.config.mjs',
+    '/.zed/',
+];
+
+/**
+ * Generated root files are untracked noise in a project whose .gitignore does
+ * not cover them. Policy consistent with the ownership model: own a fenced
+ * block, never rewrite user lines, respect a user's deletion of the block,
+ * and stand down entirely when the entries are already covered (the platform
+ * monorepo commits them itself) or after --no-gitignore.
+ */
+export function ensureGitignoreBlock(
+    context: GeneratorContext,
+    previouslyOptedOut: boolean,
+    previousState: ManifestFileState | undefined,
+    noGitignore: boolean,
+): { state: ManifestFileState; optedOut: boolean } {
+    if (noGitignore || previouslyOptedOut) {
+        return { state: 'skipped', optedOut: true };
+    }
+
+    const gitignorePath = path.join(context.projectRoot, '.gitignore');
+    const currentContent = fs.existsSync(gitignorePath) ? fs.readFileSync(gitignorePath, 'utf8') : null;
+    const hasBlock = currentContent !== null && currentContent.includes(MANAGED_BLOCK_BEGIN);
+
+    if (currentContent !== null && !hasBlock) {
+        const lines = currentContent.split(/\r?\n/).map((line) => line.trim());
+
+        if (GITIGNORE_ENTRIES.every((entry) => lines.includes(entry))) {
+            return { state: 'skipped', optedOut: false };
+        }
+
+        if (previousState === 'managed') {
+            // The user deleted the managed block — their decision stands.
+            context.instructions.push(
+                'The managed ignore block was removed from .gitignore and is left alone. ' +
+                    'Re-add it by re-running after deleting var/admin-extension-tooling/manifest.json, ' +
+                    'or silence this with --no-gitignore.',
+            );
+
+            return { state: 'skipped', optedOut: false };
+        }
+    }
+
+    const result = writeManagedBlock(
+        gitignorePath,
+        [
+            '# Generated by composer admin:setup-extension-tooling — root projections and IDE bootstraps.',
+            ...GITIGNORE_ENTRIES,
+        ],
+        context.dryRun,
+    );
+
+    record(context, result);
+
+    if (result.state === 'conflict') {
+        context.instructions.push(
+            '.gitignore contains a malformed managed block (begin marker without end) and was not touched. ' +
+                `Remove or complete the block starting with "${MANAGED_BLOCK_BEGIN}".`,
+        );
+    }
+
+    return { state: toManifestState(result.state), optedOut: false };
+}
