+++ src/Administration/Resources/app/administration/scripts/extensionTooling/setup.ts
@@ -1,7 +1,13 @@
 /**
  * @sw-package framework
  *
- * Generator for the Administration extension tooling ("Connected Toolchain").
+ * Orchestrator and CLI entrypoint for the Administration extension tooling
+ * ("Connected Toolchain"). The generation steps live in sibling modules:
+ * discovery (setup-discovery), the config generators (setup-configs), the
+ * shim/bridge cluster (setup-bridge), and the .gitignore block
+ * (setup-gitignore); all share the write-accounting context in setup-context.
+ * The previously public discovery API is re-exported here so `./setup` stays
+ * the single import for consumers.
  *
  * Pipeline: discover extensions (var/plugins.json) → per-extension leaf
  * tsconfigs in var/admin-extension-tooling/ → root tsconfig/eslint
@@ -15,69 +21,27 @@
 
 import fs from 'fs';
 import path from 'path';
-import {
-    GENERATED_MARKER,
-    MANAGED_BLOCK_BEGIN,
-    SHIM_DIR_NAME,
-    STATE_DIR,
-    asRelativeSpecifier,
-    canonicalizePath,
-    findExtensionRoot,
-    findNearestConfig,
-    isGeneratedContent,
-    isWithin,
-    readBundleConfig,
-    readPreviousManifest,
-    relativePosix,
-    toPosix,
-    writeManagedBlock,
-    writeManagedFile,
-    writeScaffoldFile,
-    writeStateFile,
-} from './shared';
 import { CliUsageError, parseCli, renderHelp } from './cli';
 import type { CommandSpec } from './cli';
-import type {
-    AdministrationTarget,
-    ExtensionToolingManifest,
-    ExtensionToolingProject,
-    ManagedFileState,
-    ManifestFileState,
-    WriteResult,
-} from './shared';
-import {
-    probeCacheEntryKey,
-    readProbeCache,
-    resolveModesFromCache,
-    resolveStaticEslintMode,
-    resolveStaticTsMode,
-    writeProbeCache,
-} from './probe';
+import { probeCacheEntryKey, readProbeCache, resolveModesFromCache, writeProbeCache } from './probe';
 import { renderSetupReport } from './report';
+import { STATE_DIR, readEslintMajorVersion, readPreviousManifest, relativePosix, writeStateFile } from './shared';
+import type { ExtensionToolingManifest, ExtensionToolingProject, WriteResult } from './shared';
+import { record, toManifestState } from './setup-context';
+import type { GeneratorContext } from './setup-context';
+import { checkDiscoveryFreshness, discoverProjects } from './setup-discovery';
+import {
+    createIdeBootstraps,
+    createLeafConfigs,
+    createRootEslintConfig,
+    createRootTsconfig,
+    ensureEntitySchema,
+} from './setup-configs';
+import { createShims } from './setup-bridge';
+import { ensureGitignoreBlock } from './setup-gitignore';
 
-const ESLINT_CONFIG_NAMES = [
-    'eslint.config.mjs',
-    'eslint.config.js',
-    'eslint.config.cjs',
-    'eslint.config.ts',
-];
-const SOURCE_EXTENSIONS = [
-    'ts',
-    'tsx',
-    'vue',
-    'js',
-];
-/**
- * Test files are split off from the runtime program (whose preset sets
- * `types: []`, so its runner globals are absent) into a dedicated spec program
- * that adds jest types. The runtime leaf excludes these suffixes; the spec leaf
- * includes exactly them.
- */
-const SPEC_FILE_SUFFIXES = [
-    'spec.ts',
-    'spec.tsx',
-    'spec.js',
-];
+// Barrel: discovery moved to setup-discovery but stays importable from ./setup.
+export { checkDiscoveryFreshness, discoverProjects };
 
 export interface SetupExtensionToolingOptions {
     projectRoot: string;
@@ -113,904 +77,62 @@ export interface SetupExtensionToolingResult {
     discoverySource: { path: string; updatedAt: string | null };
 }
 
-interface GeneratorContext {
-    projectRoot: string;
-    administrationRoot: string;
-    toolingRoot: string;
-    dryRun: boolean;
-    writes: WriteResult[];
-    staleFiles: string[];
-    warnings: string[];
-    instructions: string[];
-}
-
-function record(context: GeneratorContext, result: WriteResult): ManagedFileState {
-    // Reported project-root-relative so every consumer renders portable paths.
-    context.writes.push({
-        ...result,
-        file: path.isAbsolute(result.file) ? relativePosix(context.projectRoot, result.file) : result.file,
-    });
-
-    return result.state;
-}
-
-function toManifestState(state: ManagedFileState): ManifestFileState {
-    return state === 'conflict' || state === 'skipped' ? state : 'managed';
-}
-
-function safeFileName(name: string): string {
-    return name.replace(/[^a-zA-Z0-9_-]/g, '-').toLowerCase();
-}
-
-export function discoverProjects(
-    projectRoot: string,
-    administrationRoot: string,
-    pluginsConfigPath: string,
-): ExtensionToolingProject[] {
-    const bundles = readBundleConfig(pluginsConfigPath);
-    const administrationSourcePath = path.resolve(administrationRoot, 'src');
-    const vendorRoot = path.join(projectRoot, 'vendor');
-
-    interface ProjectGroup {
-        extensionRoot: string;
-        technicalNames: Set<string>;
-        targets: Map<string, { sourcePath: string; technicalNames: Set<string> }>;
-    }
-
-    const groups = new Map<string, ProjectGroup>();
-
-    for (const bundle of bundles) {
-        const administrationPath = bundle.administration?.path;
-
-        if (!administrationPath) {
-            continue;
-        }
-
-        const bundleBasePath = path.isAbsolute(bundle.basePath)
-            ? path.normalize(bundle.basePath)
-            : path.resolve(projectRoot, bundle.basePath);
-        const sourcePath = path.resolve(bundleBasePath, administrationPath);
-
-        if (!fs.existsSync(sourcePath) || sourcePath === administrationSourcePath) {
-            continue;
-        }
-
-        const extensionRoot = findExtensionRoot(projectRoot, bundleBasePath);
-        const group = groups.get(extensionRoot) ?? {
-            extensionRoot,
-            technicalNames: new Set<string>(),
-            targets: new Map<string, { sourcePath: string; technicalNames: Set<string> }>(),
-        };
-
-        group.technicalNames.add(bundle.technicalName);
-        const canonicalSourcePath = canonicalizePath(sourcePath);
-        const target = group.targets.get(canonicalSourcePath) ?? {
-            // Keep paths in the caller's project-root namespace for portable
-            // manifest entries; use the canonical form only as the dedupe key.
-            sourcePath: path.normalize(sourcePath),
-            technicalNames: new Set<string>(),
-        };
-
-        target.technicalNames.add(bundle.technicalName);
-        group.targets.set(canonicalSourcePath, target);
-        groups.set(extensionRoot, group);
-    }
-
-    const usedNames = new Map<string, number>();
-
-    return [...groups.values()]
-        .sort((left, right) => left.extensionRoot.localeCompare(right.extensionRoot))
-        .map((group) => {
-            const baseName = path.basename(group.extensionRoot);
-            const nameCount = usedNames.get(baseName) ?? 0;
-
-            usedNames.set(baseName, nameCount + 1);
-
-            const name = nameCount === 0 ? baseName : `${path.basename(path.dirname(group.extensionRoot))}-${baseName}`;
-            const targets: AdministrationTarget[] = [...group.targets.values()]
-                .sort((left, right) => left.sourcePath.localeCompare(right.sourcePath))
-                .map((target) => {
-                    const tsconfig = findNearestConfig(target.sourcePath, group.extensionRoot, ['tsconfig.json']);
-                    const eslintConfig = findNearestConfig(target.sourcePath, group.extensionRoot, ESLINT_CONFIG_NAMES);
-                    // A bridge sits beside the config it serves: the source root
-                    // itself (per-root shim) or the directory of the owned config
-                    // that extends it (root-config shim). Check every candidate so
-                    // a root-config bridge is not mistaken for "no bridge".
-                    const bridgeDirs = new Set<string>([path.dirname(target.sourcePath)]);
-
-                    if (tsconfig) {
-                        bridgeDirs.add(path.dirname(tsconfig));
-                    }
-
-                    if (eslintConfig) {
-                        bridgeDirs.add(path.dirname(eslintConfig));
-                    }
-
-                    return {
-                        technicalNames: [...target.technicalNames].sort(),
-                        sourcePath: relativePosix(projectRoot, target.sourcePath),
-                        adminFolder: relativePosix(projectRoot, path.dirname(target.sourcePath)),
-                        bridgePresent: [...bridgeDirs].some((dir) =>
-                            fs.existsSync(path.join(dir, SHIM_DIR_NAME, 'tsconfig.json')),
-                        ),
-                        tsconfig: tsconfig ? relativePosix(projectRoot, tsconfig) : null,
-                        eslintConfig: eslintConfig ? relativePosix(projectRoot, eslintConfig) : null,
-                        ts: resolveStaticTsMode(tsconfig),
-                        eslint: resolveStaticEslintMode(eslintConfig),
-                        checkTsconfig: '',
-                        specTsconfig: '',
-                    };
-                });
-
-            return {
-                name,
-                technicalNames: [...group.technicalNames].sort(),
-                basePath: relativePosix(projectRoot, group.extensionRoot),
-                vendor: isWithin(group.extensionRoot, vendorRoot),
-                targets,
-            };
-        });
-}
-
-/**
- * Discovery reads var/plugins.json, which only `bin/console bundle:dump`
- * refreshes — neither plugin:install nor cache:clear do. A freshly activated
- * plugin is invisible until then, so a stale file earns a hint instead of a
- * silently green "up to date". Heuristic: false positives cost one dim line.
- */
-export function checkDiscoveryFreshness(projectRoot: string, pluginsConfigPath: string): string | null {
-    try {
-        const pluginsMtime = fs.statSync(pluginsConfigPath).mtimeMs;
-        const customPluginsDir = path.join(projectRoot, 'custom', 'plugins');
-        let newestPluginMtime = 0;
-
-        for (const entry of fs.readdirSync(customPluginsDir, { withFileTypes: true })) {
-            if (!entry.isDirectory()) {
-                continue;
-            }
-
-            try {
-                newestPluginMtime = Math.max(
-                    newestPluginMtime,
-                    fs.statSync(path.join(customPluginsDir, entry.name, 'composer.json')).mtimeMs,
-                );
-            } catch {
-                // A plugin folder without composer.json cannot be discovered anyway.
-            }
-        }
-
-        if (newestPluginMtime > pluginsMtime) {
-            return (
-                'var/plugins.json is older than custom/plugins/ — if an extension is missing below, ' +
-                'run: bin/console bundle:dump'
-            );
-        }
-    } catch {
-        // Missing plugins.json or custom/plugins/ — discovery itself reports that.
-    }
-
-    return null;
-}
-
-function ensureEntitySchema(context: GeneratorContext): boolean {
-    const entitySchemaPath = path.join(context.administrationRoot, 'src', 'entity-schema-definition.d.ts');
-
-    if (fs.existsSync(entitySchemaPath)) {
-        return !isGeneratedContent(fs.readFileSync(entitySchemaPath, 'utf8'));
-    }
-
-    const stubContent = [
-        '/* eslint-disable */',
-        `/* ${GENERATED_MARKER} — STUB.`,
-        '   The installation-specific entity schema has not been generated yet, so',
-        '   `EntitySchema.Entities` is empty and every entity name fails the type',
-        '   check instead of silently degrading to `any`.',
-        '   Generate the real file with: composer admin:generate-entity-schema-types */',
-        'declare namespace EntitySchema {',
-        '    interface Entities {}',
-        '}',
-        '',
-    ].join('\n');
-
-    record(context, writeManagedFile(entitySchemaPath, stubContent, context.dryRun));
-    context.warnings.push(
-        'The generated entity schema types are missing; a stub with an empty entity list was created. ' +
-            'Run `composer admin:generate-entity-schema-types` to get installation-specific entity types.',
-    );
-
-    return false;
-}
-
-function createLeafConfigs(context: GeneratorContext, projects: ExtensionToolingProject[]): ExtensionToolingProject[] {
-    const projectsDir = path.join(context.projectRoot, STATE_DIR, 'projects');
-    const basePreset = path.join(context.administrationRoot, 'extension-tooling', 'tsconfig.base.json');
-    const adminTypes = path.join(context.administrationRoot, 'extension-tooling', 'admin-types.d.ts');
-    const specTypes = path.join(context.administrationRoot, 'extension-tooling', 'spec-types.d.ts');
-    const usedFileNames = new Set<string>();
-
-    const sourceGlobs = (configPath: string, sourcePath: string, extensions: string[]): string[] =>
-        extensions.map(
-            (extension) =>
-                `${asRelativeSpecifier(configPath, path.resolve(context.projectRoot, sourcePath))}/**/*.${extension}`,
-        );
-
-    const configured = projects.map((project) => ({
-        ...project,
-        targets: project.targets.map((target) => {
-            const targetName =
-                project.targets.length === 1
-                    ? project.name
-                    : `${project.name}-${target.technicalNames[0] ?? path.basename(target.adminFolder)}`;
-            let fileName = `${safeFileName(targetName)}.json`;
-            let suffix = 2;
-
-            while (usedFileNames.has(fileName)) {
-                fileName = `${safeFileName(targetName)}-${suffix}.json`;
-                suffix += 1;
-            }
-
-            const specFileName = fileName.replace(/\.json$/, '-specs.json');
-
-            usedFileNames.add(fileName);
-            usedFileNames.add(specFileName);
-
-            const configPath = path.join(projectsDir, fileName);
-            const content = `// ${GENERATED_MARKER}\n${JSON.stringify(
-                {
-                    extends: asRelativeSpecifier(configPath, basePreset),
-                    files: [asRelativeSpecifier(configPath, adminTypes)],
-                    include: sourceGlobs(configPath, target.sourcePath, SOURCE_EXTENSIONS),
-                    // Exclude patterns resolve relative to this config, so they
-                    // carry the same source prefix as the includes.
-                    exclude: sourceGlobs(configPath, target.sourcePath, SPEC_FILE_SUFFIXES),
-                },
-                null,
-                4,
-            )}\n`;
-
-            record(context, writeStateFile(configPath, content, context.dryRun));
-
-            // Companion program for spec files: the same type surface plus jest
-            // types (spec-types.d.ts), including only this target's specs.
-            const specConfigPath = path.join(projectsDir, specFileName);
-            const specContent = `// ${GENERATED_MARKER}\n${JSON.stringify(
-                {
-                    extends: asRelativeSpecifier(specConfigPath, basePreset),
-                    // Point typeRoots at the Administration's own @types so the jest
-                    // reference in spec-types.d.ts resolves — a triple-slash
-                    // reference is looked up relative to this config, not the .d.ts.
-                    compilerOptions: {
-                        typeRoots: [
-                            asRelativeSpecifier(
-                                specConfigPath,
-                                path.join(context.administrationRoot, 'node_modules', '@types'),
-                            ),
-                        ],
-                    },
-                    files: [
-                        asRelativeSpecifier(specConfigPath, adminTypes),
-                        asRelativeSpecifier(specConfigPath, specTypes),
-                    ],
-                    include: sourceGlobs(specConfigPath, target.sourcePath, SPEC_FILE_SUFFIXES),
-                },
-                null,
-                4,
-            )}\n`;
-
-            record(context, writeStateFile(specConfigPath, specContent, context.dryRun));
-
-            return {
-                ...target,
-                checkTsconfig: target.tsconfig ?? relativePosix(context.projectRoot, configPath),
-                specTsconfig: relativePosix(context.projectRoot, specConfigPath),
-            };
-        }),
-    }));
-
-    if (fs.existsSync(projectsDir)) {
-        for (const existingFile of fs.readdirSync(projectsDir)) {
-            if (existingFile.endsWith('.json') && !usedFileNames.has(existingFile)) {
-                context.staleFiles.push(relativePosix(context.projectRoot, path.join(projectsDir, existingFile)));
-
-                if (!context.dryRun) {
-                    fs.rmSync(path.join(projectsDir, existingFile));
-                }
-            }
-        }
-    }
-
-    return configured;
-}
-
-function createRootTsconfig(context: GeneratorContext, projects: ExtensionToolingProject[]): ManagedFileState {
-    const rootTsconfigPath = path.join(context.projectRoot, 'tsconfig.json');
-    // Reference both the runtime and spec leaves so the IDE — and ESLint's
-    // project service — can associate every managed source and spec file with a
-    // program (this is what enables type-aware linting of managed specs).
-    const references = projects.flatMap((project) =>
-        project.targets
-            .filter((target) => target.ts.mode === 'managed')
-            .flatMap((target) => [
-                { path: `./${target.checkTsconfig}` },
-                { path: `./${target.specTsconfig}` },
-            ]),
-    );
-    const content = `// ${GENERATED_MARKER} — solution-style index routing each extension file to its leaf project.\n${JSON.stringify(
-        {
-            files: [],
-            references,
-        },
-        null,
-        4,
-    )}\n`;
-    const state = record(context, writeManagedFile(rootTsconfigPath, content, context.dryRun));
-
-    if (state === 'conflict') {
-        context.instructions.push(
-            [
-                `${rootTsconfigPath} exists and is not managed by this tool. To integrate, add these references:`,
-                ...references.map((reference) => `    { "path": "${reference.path}" }`),
-                'or remove the file and re-run `composer admin:setup-extension-tooling`.',
-            ].join('\n'),
-        );
-    }
-
-    return state;
-}
-
-function createRootEslintConfig(context: GeneratorContext, projects: ExtensionToolingProject[]): ManagedFileState {
-    const rootEslintPath = path.join(context.projectRoot, 'eslint.config.mjs');
-    const extensionRoots = [...new Set(projects.flatMap((project) => project.targets.map((target) => target.sourcePath)))];
-
-    // Without discovered extensions there is nothing to scope the config to.
-    // The factory would then fall back to project-wide globs and lint files
-    // outside the Administration (e.g. real server-side Twig), so the root
-    // config is skipped entirely instead of written empty.
-    if (extensionRoots.length === 0) {
-        return 'skipped';
-    }
-
-    const factoryPath = path.join(context.administrationRoot, 'extension-tooling', 'eslint.mjs');
-    const factorySpecifier = asRelativeSpecifier(rootEslintPath, factoryPath);
-    const content = [
-        `// ${GENERATED_MARKER}`,
-        `import { shopwareAdminExtension } from ${JSON.stringify(factorySpecifier)};`,
-        '',
-        'export default shopwareAdminExtension({',
-        '    tsconfigRootDir: import.meta.dirname,',
-        '    // Spec files are type-aware-linted here — their leaves are referenced from the root tsconfig.',
-        '    typedSpecs: true,',
-        '    extensionRoots: [',
-        ...extensionRoots.map((extensionRoot) => `        ${JSON.stringify(extensionRoot)},`),
-        '    ],',
-        '});',
-        '',
-    ].join('\n');
-    const state = record(context, writeManagedFile(rootEslintPath, content, context.dryRun));
-
-    if (state === 'conflict') {
-        context.instructions.push(
-            [
-                `${rootEslintPath} exists and is not managed by this tool. To integrate, compose the shared factory:`,
-                `    import { shopwareAdminExtension } from ${JSON.stringify(factorySpecifier)};`,
-                '    export default [',
-                '        ...shopwareAdminExtension({ tsconfigRootDir: import.meta.dirname, extensionRoots: [/* … */] }),',
-                '        // your own config',
-                '    ];',
-            ].join('\n'),
-        );
-    }
-
-    return state;
-}
-
-function readEslintMajorVersion(administrationRoot: string): number {
-    const eslintPackagePath = path.join(administrationRoot, 'node_modules', 'eslint', 'package.json');
-
-    try {
-        const eslintPackage = JSON.parse(fs.readFileSync(eslintPackagePath, 'utf8')) as { version: string };
-
-        return Number(eslintPackage.version.split('.')[0]);
-    } catch {
-        return 9;
-    }
-}
-
-function createIdeBootstraps(context: GeneratorContext, adminRelative: string): Record<string, ManagedFileState> {
-    const states: Record<string, ManagedFileState> = {};
-    const eslintFlags = readEslintMajorVersion(context.administrationRoot) < 10 ? ['v10_config_lookup_from_file'] : [];
-    const bootstraps: Array<{ key: string; file: string; content: string; settings: string[] }> = [
-        {
-            key: '.vscode/settings.json',
-            file: path.join(context.projectRoot, '.vscode', 'settings.json'),
-            content: `// ${GENERATED_MARKER}\n${JSON.stringify(
-                {
-                    'typescript.tsdk': `${adminRelative}/node_modules/typescript/lib`,
-                    'eslint.nodePath': `${adminRelative}/node_modules`,
-                    ...(eslintFlags.length > 0 ? { 'eslint.options': { flags: eslintFlags } } : {}),
-                    'eslint.validate': [
-                        'javascript',
-                        'typescript',
-                        'vue',
-                        'twig',
-                    ],
-                    'files.associations': { '*.html.twig': 'twig' },
-                },
-                null,
-                4,
-            )}\n`,
-            settings: [
-                `"typescript.tsdk": "${adminRelative}/node_modules/typescript/lib"`,
-                `"eslint.nodePath": "${adminRelative}/node_modules"`,
-                ...(eslintFlags.length > 0 ? [`"eslint.options": { "flags": ["${eslintFlags[0]}"] }`] : []),
-            ],
-        },
-        {
-            key: '.zed/settings.json',
-            file: path.join(context.projectRoot, '.zed', 'settings.json'),
-            content: `// ${GENERATED_MARKER}\n${JSON.stringify(
-                {
-                    lsp: {
-                        vtsls: {
-                            initialization_options: {
-                                typescript: { tsdk: `${adminRelative}/node_modules/typescript/lib` },
-                            },
-                        },
-                        eslint: {
-                            settings: {
-                                nodePath: `${adminRelative}/node_modules`,
-                                ...(eslintFlags.length > 0 ? { options: { flags: eslintFlags } } : {}),
-                            },
-                        },
-                    },
-                },
-                null,
-                4,
-            )}\n`,
-            settings: [
-                `"lsp.vtsls.initialization_options.typescript.tsdk": "${adminRelative}/node_modules/typescript/lib"`,
-                `"lsp.eslint.settings.nodePath": "${adminRelative}/node_modules"`,
-            ],
-        },
-    ];
-
-    for (const bootstrap of bootstraps) {
-        if (fs.existsSync(bootstrap.file) && !isGeneratedContent(fs.readFileSync(bootstrap.file, 'utf8'))) {
-            states[bootstrap.key] = 'skipped';
-            record(context, { file: bootstrap.file, state: 'skipped' });
-            context.instructions.push(
-                [
-                    `${bootstrap.file} is user-owned and was not touched. For IDE support add:`,
-                    ...bootstrap.settings.map((setting) => `    ${setting}`),
-                ].join('\n'),
-            );
-
-            continue;
-        }
-
-        states[bootstrap.key] = record(context, writeManagedFile(bootstrap.file, bootstrap.content, context.dryRun));
-    }
-
-    // PhpStorm stores these settings in .idea internals we never write to.
-    states['.idea'] = 'skipped';
-    context.instructions.push(
-        [
-            'PhpStorm (configure once, Settings → Languages & Frameworks):',
-            `    TypeScript → set "TypeScript" package to ${adminRelative}/node_modules/typescript`,
-            '    JavaScript → Code Quality Tools → ESLint → Manual configuration:',
-            `        ESLint package: ${adminRelative}/node_modules/eslint`,
-            ...(readEslintMajorVersion(context.administrationRoot) < 10
-                ? ['        Extra eslint options: --flag v10_config_lookup_from_file']
-                : []),
-        ].join('\n'),
-    );
-
-    return states;
-}
-
-const GITIGNORE_ENTRIES = [
-    '/tsconfig.json',
-    '/eslint.config.mjs',
-    '/.zed/',
-];
-
-/**
- * Generated root files are untracked noise in a project whose .gitignore does
- * not cover them. Policy consistent with the ownership model: own a fenced
- * block, never rewrite user lines, respect a user's deletion of the block,
- * and stand down entirely when the entries are already covered (the platform
- * monorepo commits them itself) or after --no-gitignore.
- */
-function ensureGitignoreBlock(
-    context: GeneratorContext,
-    previouslyOptedOut: boolean,
-    previousState: ManifestFileState | undefined,
-    noGitignore: boolean,
-): { state: ManifestFileState; optedOut: boolean } {
-    if (noGitignore || previouslyOptedOut) {
-        return { state: 'skipped', optedOut: true };
-    }
-
-    const gitignorePath = path.join(context.projectRoot, '.gitignore');
-    const currentContent = fs.existsSync(gitignorePath) ? fs.readFileSync(gitignorePath, 'utf8') : null;
-    const hasBlock = currentContent !== null && currentContent.includes(MANAGED_BLOCK_BEGIN);
-
-    if (currentContent !== null && !hasBlock) {
-        const lines = currentContent.split(/\r?\n/).map((line) => line.trim());
-
-        if (GITIGNORE_ENTRIES.every((entry) => lines.includes(entry))) {
-            return { state: 'skipped', optedOut: false };
-        }
-
-        if (previousState === 'managed') {
-            // The user deleted the managed block — their decision stands.
-            context.instructions.push(
-                'The managed ignore block was removed from .gitignore and is left alone. ' +
-                    'Re-add it by re-running after deleting var/admin-extension-tooling/manifest.json, ' +
-                    'or silence this with --no-gitignore.',
-            );
-
-            return { state: 'skipped', optedOut: false };
-        }
-    }
-
-    const result = writeManagedBlock(
-        gitignorePath,
-        [
-            '# Generated by composer admin:setup-extension-tooling — root projections and IDE bootstraps.',
-            ...GITIGNORE_ENTRIES,
-        ],
-        context.dryRun,
-    );
-
-    record(context, result);
-
-    if (result.state === 'conflict') {
-        context.instructions.push(
-            '.gitignore contains a malformed managed block (begin marker without end) and was not touched. ' +
-                `Remove or complete the block starting with "${MANAGED_BLOCK_BEGIN}".`,
-        );
-    }
-
-    return { state: toManifestState(result.state), optedOut: false };
-}
-
-/**
- * Aliases contributing to a bridge's merged `paths`. Each source's targets
- * resolve relative to its own `baseDir`, so a root-config bridge can carry the
- * aliases declared beside every covered Administration root, not only one.
- */
-interface AliasSource {
-    aliasesPath: string;
-    baseDir: string;
-}
-
-function dedupeAliasSources(sources: AliasSource[]): AliasSource[] {
-    const seen = new Set<string>();
-
-    return sources.filter((source) => {
-        if (seen.has(source.aliasesPath)) {
-            return false;
-        }
-
-        seen.add(source.aliasesPath);
-
-        return true;
-    });
-}
-
-/**
- * TypeScript replaces `paths` wholesale across `extends`, so a plugin that
- * declares its own aliases would erase the preset's host paths. The bridge is
- * therefore the single `paths` declarer: it merges the preset's host paths
- * (re-relativized to the bridge, machine paths are fine in generated files)
- * with plugin aliases from every `tsconfig.aliases.json` beside a covered root
- * ({ "MyPlugin/*": ["src/*"] }, targets relative to that alias file's directory).
- */
-function buildShimPaths(
-    context: GeneratorContext,
-    shimDir: string,
-    aliasSources: AliasSource[],
-): Record<string, string[]> | null {
-    const presentSources = aliasSources.filter((source) => fs.existsSync(source.aliasesPath));
-
-    if (presentSources.length === 0) {
-        return null;
-    }
-
-    const shimTsconfigPath = path.join(shimDir, 'tsconfig.json');
-    const presetPath = path.join(context.toolingRoot, 'tsconfig.base.json');
-    const preset = JSON.parse(fs.readFileSync(presetPath, 'utf8')) as {
-        compilerOptions?: { paths?: Record<string, string[]> };
-    };
-    const presetDir = path.dirname(presetPath);
-    const mergedPaths: Record<string, string[]> = {};
+/** Reads the host-module map and warns about any declared module missing from the installed Administration. */
+function loadHostModules(context: GeneratorContext): Record<string, string> {
+    const hostModulesPath = path.join(context.toolingRoot, 'host-modules.json');
+    const hostModules = (JSON.parse(fs.readFileSync(hostModulesPath, 'utf8')) as { hostModules: Record<string, string> })
+        .hostModules;
 
     for (const [
         moduleName,
-        targets,
-    ] of Object.entries(preset.compilerOptions?.paths ?? {})) {
-        mergedPaths[moduleName] = targets.map((target) =>
-            asRelativeSpecifier(shimTsconfigPath, path.resolve(presetDir, target)),
-        );
-    }
-
-    for (const source of presentSources) {
-        const aliases = JSON.parse(fs.readFileSync(source.aliasesPath, 'utf8')) as Record<string, string[] | string>;
-
-        for (const [
-            alias,
-            targets,
-        ] of Object.entries(aliases)) {
-            mergedPaths[alias] = (Array.isArray(targets) ? targets : [targets]).map((target) =>
-                asRelativeSpecifier(shimTsconfigPath, path.resolve(source.baseDir, target)),
+        modulePath,
+    ] of Object.entries(hostModules)) {
+        if (!fs.existsSync(path.join(context.administrationRoot, modulePath))) {
+            context.warnings.push(
+                `Host module "${moduleName}" is declared in host-modules.json but ${modulePath} does not exist ` +
+                    'in the installed Administration.',
             );
         }
     }
 
-    return mergedPaths;
+    return hostModules;
 }
 
 /**
- * Writes the three git-ignored bridge files into `<bridgeParent>/.shopware-admin/`.
- * The bridge is the machine-specific hop into the installed Administration; the
- * eslint bridge derives its own parent directory at runtime, so one bridge serves
- * whichever config — per-root or a shared package root — sits beside it.
+ * Adopt verified verdicts from earlier check runs where the config inputs are
+ * unchanged, and prune cache entries of extensions that have disappeared.
  */
-function writeBridge(context: GeneratorContext, bridgeParent: string, aliasSources: AliasSource[]): void {
-    const shimDir = path.join(bridgeParent, SHIM_DIR_NAME);
-    const basePreset = path.join(context.administrationRoot, 'extension-tooling', 'tsconfig.base.json');
-    const adminTypes = path.join(context.administrationRoot, 'extension-tooling', 'admin-types.d.ts');
-    const factoryPath = path.join(context.administrationRoot, 'extension-tooling', 'eslint.mjs');
-    const shimTsconfigPath = path.join(shimDir, 'tsconfig.json');
-    const shimEslintPath = path.join(shimDir, 'eslint.mjs');
-    const shimPaths = buildShimPaths(context, shimDir, aliasSources);
+function applyProbeCache(context: GeneratorContext, discovered: ExtensionToolingProject[]): ExtensionToolingProject[] {
+    const probeCache = readProbeCache(context.projectRoot);
 
-    record(context, writeManagedFile(path.join(shimDir, '.gitignore'), `# ${GENERATED_MARKER}\n*\n`, context.dryRun));
-    record(
-        context,
-        writeManagedFile(
-            shimTsconfigPath,
-            `// ${GENERATED_MARKER} — machine-specific path into the installed Administration.\n${JSON.stringify(
-                {
-                    extends: asRelativeSpecifier(shimTsconfigPath, basePreset),
-                    files: [asRelativeSpecifier(shimTsconfigPath, adminTypes)],
-                    ...(shimPaths ? { compilerOptions: { paths: shimPaths } } : {}),
-                },
-                null,
-                4,
-            )}\n`,
-            context.dryRun,
-        ),
+    if (!probeCache) {
+        return discovered;
+    }
+
+    const withCachedModes = discovered.map((project) => ({
+        ...project,
+        targets: project.targets.map((target) => ({
+            ...target,
+            ...resolveModesFromCache(
+                target,
+                probeCacheEntryKey(project.name, target),
+                probeCache,
+                context.projectRoot,
+                context.administrationRoot,
+            ),
+        })),
+    }));
+
+    const knownNames = new Set(
+        withCachedModes.flatMap((project) => project.targets.map((target) => probeCacheEntryKey(project.name, target))),
     );
-    record(
-        context,
-        writeManagedFile(
-            shimEslintPath,
-            [
-                `// ${GENERATED_MARKER} — machine-specific path into the installed Administration.`,
-                "import path from 'node:path';",
-                "import { fileURLToPath } from 'node:url';",
-                `import { shopwareAdminExtension } from ${JSON.stringify(asRelativeSpecifier(shimEslintPath, factoryPath))};`,
-                '',
-                'const adminFolder = path.dirname(path.dirname(fileURLToPath(import.meta.url)));',
-                '',
-                `export * from ${JSON.stringify(asRelativeSpecifier(shimEslintPath, factoryPath))};`,
-                '',
-                'export function shopwareAdminExtensionConfig(options = {}) {',
-                '    return shopwareAdminExtension({ tsconfigRootDir: adminFolder, ...options });',
-                '}',
-                '',
-                'export default shopwareAdminExtensionConfig();',
-                '',
-            ].join('\n'),
-            context.dryRun,
-        ),
-    );
-}
+    const prunedEntries = Object.fromEntries(Object.entries(probeCache.entries).filter(([name]) => knownNames.has(name)));
 
-type BridgePlan = { kind: 'per-root' } | { kind: 'root-config'; dir: string } | { kind: 'ambiguous'; candidates: string[] };
-
-/**
- * Decides whether an extension is bridged once beside a package-level config
- * that already governs several Administration roots, or once per root. An
- * explicit `--root-config` always wins. Otherwise a single owned config shared
- * by two or more roots is adopted as the root config; two or more competing
- * shared configs are ambiguous and require the explicit flag; anything else (a
- * single root, genuinely independent per-root configs, or zero-config) stays
- * per-root, so independent layouts keep their existing behavior.
- */
-function resolveBridgePlan(
-    context: GeneratorContext,
-    project: ExtensionToolingProject,
-    explicitRootConfig?: string,
-): BridgePlan {
-    const extensionRoot = path.resolve(context.projectRoot, project.basePath);
-
-    if (explicitRootConfig !== undefined) {
-        return { kind: 'root-config', dir: path.resolve(extensionRoot, explicitRootConfig) };
+    if (!context.dryRun && Object.keys(prunedEntries).length !== Object.keys(probeCache.entries).length) {
+        writeProbeCache(context.projectRoot, { version: 2, entries: prunedEntries });
     }
 
-    if (project.targets.length < 2) {
-        return { kind: 'per-root' };
-    }
-
-    const sharedConfigDirs = (configOf: (target: AdministrationTarget) => string | null): string[] => {
-        const counts = new Map<string, number>();
-
-        for (const target of project.targets) {
-            const config = configOf(target);
-
-            if (config) {
-                const dir = path.dirname(path.resolve(context.projectRoot, config));
-
-                counts.set(dir, (counts.get(dir) ?? 0) + 1);
-            }
-        }
-
-        return [...counts.entries()]
-            .filter(
-                ([
-                    ,
-                    count,
-                ]) => count >= 2,
-            )
-            .map(([dir]) => dir);
-    };
-    const candidates = [
-        ...new Set([
-            ...sharedConfigDirs((target) => target.tsconfig),
-            ...sharedConfigDirs((target) => target.eslintConfig),
-        ]),
-    ];
-
-    if (candidates.length === 0) {
-        return { kind: 'per-root' };
-    }
-
-    if (candidates.length === 1) {
-        return { kind: 'root-config', dir: candidates[0] };
-    }
-
-    return { kind: 'ambiguous', candidates };
-}
-
-function createShims(
-    context: GeneratorContext,
-    projects: ExtensionToolingProject[],
-    shim: string,
-    rootConfig?: string,
-): void {
-    const selected =
-        shim === 'all-custom'
-            ? projects.filter((project) => project.basePath.startsWith('custom/plugins/'))
-            : projects.filter((project) => project.name === shim || project.technicalNames.includes(shim));
-
-    if (selected.length === 0) {
-        const available = projects.map((project) => project.name).join(', ');
-
-        throw new Error(`No extension matches --shim=${shim}. Discovered extensions: ${available || '(none)'}.`);
-    }
-
-    if (rootConfig !== undefined && shim === 'all-custom') {
-        throw new Error('--root-config cannot be combined with --shim=all-custom; bridge one extension at a time.');
-    }
-
-    for (const project of selected) {
-        if (!project.basePath.startsWith('custom/plugins/')) {
-            throw new Error(
-                `Refusing to write a shim into ${project.basePath}: shims are only generated below custom/plugins ` +
-                    '(vendor and platform extensions are checked through host-owned configs instead).',
-            );
-        }
-
-        const plan = resolveBridgePlan(context, project, rootConfig);
-
-        if (plan.kind === 'ambiguous') {
-            const relative = plan.candidates
-                .map((dir) => relativePosix(context.projectRoot, dir))
-                .sort()
-                .join(', ');
-
-            throw new Error(
-                `${project.name} has more than one package-level config governing multiple Administration roots ` +
-                    `(${relative}). Choose one with --root-config=<dir> relative to ${project.basePath}, ` +
-                    'or give each root its own independent config.',
-            );
-        }
-
-        if (plan.kind === 'root-config') {
-            const coveredAdminFolders = project.targets.map((target) =>
-                path.resolve(context.projectRoot, target.adminFolder),
-            );
-            const aliasSources = dedupeAliasSources([
-                { aliasesPath: path.join(plan.dir, 'tsconfig.aliases.json'), baseDir: plan.dir },
-                ...coveredAdminFolders.map((folder) => ({
-                    aliasesPath: path.join(folder, 'tsconfig.aliases.json'),
-                    baseDir: folder,
-                })),
-            ]);
-
-            writeBridge(context, plan.dir, aliasSources);
-            scaffoldExtensionConfigs(
-                context,
-                project.name,
-                plan.dir,
-                project.targets.map((target) => target.sourcePath),
-            );
-
-            continue;
-        }
-
-        for (const target of project.targets) {
-            const adminFolder = path.resolve(context.projectRoot, target.adminFolder);
-
-            writeBridge(context, adminFolder, [
-                { aliasesPath: path.join(adminFolder, 'tsconfig.aliases.json'), baseDir: adminFolder },
-            ]);
-            scaffoldExtensionConfigs(context, project.name, adminFolder, [target.sourcePath]);
-        }
-    }
-}
-
-/**
- * Scaffolds the plugin's own small, committable tsconfig/eslint that extend the
- * generated `.shopware-admin/` bridge in `configDir` — created only when absent
- * so the developer can see and customize them. In root-config mode one pair
- * includes every source root; per root, one pair covers a single root. An
- * existing config is never overwritten; instead we print the one line to add so
- * it too composes the preset.
- */
-function scaffoldExtensionConfigs(context: GeneratorContext, name: string, configDir: string, sourcePaths: string[]): void {
-    const include = sourcePaths.flatMap((sourcePath) => {
-        const sourceRelative = toPosix(path.relative(configDir, path.resolve(context.projectRoot, sourcePath)));
-
-        return SOURCE_EXTENSIONS.map((extension) => `${sourceRelative}/**/*.${extension}`);
-    });
-    const tsconfigPath = path.join(configDir, 'tsconfig.json');
-    const eslintPath = path.join(configDir, 'eslint.config.mjs');
-    const tsconfigContent =
-        `// Committed config for ${name}. Extends the generated Shopware bridge in .shopware-admin/\n` +
-        '// (git-ignored, holds the machine-specific paths). Safe to edit and commit — keep the "extends".\n' +
-        '// Spec files stay excluded here — the check type-checks them separately with jest types.\n' +
-        `${JSON.stringify(
-            {
-                extends: './.shopware-admin/tsconfig.json',
-                include,
-                exclude: SPEC_FILE_SUFFIXES.map((suffix) => `**/*.${suffix}`),
-            },
-            null,
-            4,
-        )}\n`;
-    const eslintContent = [
-        `// Committed config for ${name}. Composes the generated Shopware bridge in .shopware-admin/`,
-        '// (git-ignored). Safe to edit and commit — keep the import and the ...spread.',
-        "import shopware from './.shopware-admin/eslint.mjs';",
-        '',
-        'export default [',
-        '    ...shopware,',
-        '    // Add your own rules here.',
-        '];',
-        '',
-    ].join('\n');
-
-    const tsconfigResult = record(context, writeScaffoldFile(tsconfigPath, tsconfigContent, context.dryRun));
-    const eslintResult = record(context, writeScaffoldFile(eslintPath, eslintContent, context.dryRun));
-
-    if (tsconfigResult === 'skipped') {
-        context.warnings.push(
-            `${tsconfigPath} already exists and was not touched — add \`"extends": "./.shopware-admin/tsconfig.json"\` ` +
-                `so ${name} composes the Shopware preset. Own "files" array? Remove it — the bridge provides the ` +
-                'type surface. Own paths? Declare them in tsconfig.aliases.json.',
-        );
-    }
-
-    if (eslintResult === 'skipped') {
-        context.warnings.push(
-            `${eslintPath} already exists and was not touched — compose the bridge in it: ` +
-                "import shopware from './.shopware-admin/eslint.mjs'; export default [ ...shopware /* , your rules */ ];",
-        );
-    }
+    return withCachedModes;
 }
 
 export function setupExtensionTooling(options: SetupExtensionToolingOptions): SetupExtensionToolingResult {
@@ -1028,22 +150,7 @@ export function setupExtensionTooling(options: SetupExtensionToolingOptions): Se
         instructions: [],
     };
 
-    const hostModulesPath = path.join(context.toolingRoot, 'host-modules.json');
-    const hostModules = (JSON.parse(fs.readFileSync(hostModulesPath, 'utf8')) as { hostModules: Record<string, string> })
-        .hostModules;
-
-    for (const [
-        moduleName,
-        modulePath,
-    ] of Object.entries(hostModules)) {
-        if (!fs.existsSync(path.join(administrationRoot, modulePath))) {
-            context.warnings.push(
-                `Host module "${moduleName}" is declared in host-modules.json but ${modulePath} does not exist ` +
-                    'in the installed Administration.',
-            );
-        }
-    }
-
+    const hostModules = loadHostModules(context);
     const entitySchemaAvailable = ensureEntitySchema(context);
     const freshnessWarning = checkDiscoveryFreshness(projectRoot, pluginsConfigPath);
 
@@ -1060,42 +167,13 @@ export function setupExtensionTooling(options: SetupExtensionToolingOptions): Se
         discovered = discoverProjects(projectRoot, administrationRoot, pluginsConfigPath);
     }
 
-    // Adopt verified verdicts from earlier check runs where the config inputs
-    // are unchanged; prune cache entries of extensions that disappeared.
-    const probeCache = readProbeCache(projectRoot);
-
-    if (probeCache) {
-        discovered = discovered.map((project) => ({
-            ...project,
-            targets: project.targets.map((target) => ({
-                ...target,
-                ...resolveModesFromCache(
-                    target,
-                    probeCacheEntryKey(project.name, target),
-                    probeCache,
-                    projectRoot,
-                    administrationRoot,
-                ),
-            })),
-        }));
-
-        const knownNames = new Set(
-            discovered.flatMap((project) => project.targets.map((target) => probeCacheEntryKey(project.name, target))),
-        );
-        const prunedEntries = Object.fromEntries(
-            Object.entries(probeCache.entries).filter(([name]) => knownNames.has(name)),
-        );
-
-        if (!context.dryRun && Object.keys(prunedEntries).length !== Object.keys(probeCache.entries).length) {
-            writeProbeCache(projectRoot, { version: 2, entries: prunedEntries });
-        }
-    }
+    discovered = applyProbeCache(context, discovered);
 
     const projects = createLeafConfigs(context, discovered);
     const rootTsconfigState = createRootTsconfig(context, projects);
     const rootEslintState = createRootEslintConfig(context, projects);
     const adminRelative = relativePosix(projectRoot, administrationRoot);
-    const ideBootstraps = createIdeBootstraps(context, adminRelative);
+    const ideBootstraps = createIdeBootstraps(context, adminRelative, readEslintMajorVersion(administrationRoot));
     const previousManifest = readPreviousManifest(projectRoot);
     const gitignore = ensureGitignoreBlock(
         context,
